moved task_context_factory fixture to app/test_task.py
